Add ContainerRegistryTemplateProvider
